api/requirements.txt
# FastAPI and server
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
python-multipart>=0.0.6

# Authentication
fastapi-azure-auth>=4.3.0,<5.0.0
python-jose[cryptography]>=3.3.0

# Database
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.13.0,<2.0.0
pyodbc>=5.0.0  # For Azure SQL
aiosqlite>=0.19.0  # For SQLite async

# Utilities
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
